requirements.txt
# Configurations
hydra-core==1.4.0.dev1
hydra-joblib-launcher==1.2.0
hydra-nevergrad-sweeper==1.3.0.dev0
setuptools

# Algorithms
numpy
scipy
scikit_learn

# Data Handling
h5py
mne
lspopt
openpyxl

# Visualization
matplotlib
tabulate
seaborn
jupyter

# Install pytorch from https://pytorch.org/get-started/locally/
torch

